rename firebase-utils as altair-api-utils
