+++ packages/altair-api-utils/src/constants.ts
