+++ packages/altair-api/src/team-memberships/team-memberships.service.ts
@@ -1,7 +1,7 @@
 import {
   CreateTeamMembershipDto,
   UpdateTeamMembershipDto,
-} from '@altairgraphql/firebase-utils';
+} from '@altairgraphql/api-utils';
 import {
   BadRequestException,
   ForbiddenException,
