+++ packages/altair-dashboard/package.json
@@ -10,7 +10,8 @@
     "prepare": "yarn build"
   },
   "dependencies": {
-    "@altairgraphql/firebase-utils": "^5.0.14",
+    "@altairgraphql/api": "^5.0.14",
+    "@altairgraphql/api-utils": "^5.0.14",
     "@emotion/react": "^11.10.5",
     "@emotion/server": "^11.10.0",
     "@mantine/core": "^5.9.5",
