+++ packages/login-redirect/package.json
@@ -17,7 +17,7 @@
     "prepare": "yarn build"
   },
   "dependencies": {
-    "@altairgraphql/firebase-utils": "^5.0.14",
+    "@altairgraphql/api-utils": "^5.0.14",
     "firebase": "^9.9.2"
   },
   "private": true
